requirements.txt
streamlit
requests
PyGithub
python-dotenv
